86cyn7391 - file restructure to have each role dashboard in one module
