--- app/client/ui-modules/agent-dashboard/components/ProgressBar.tsx
@@ -1,24 +1,0 @@
-"use client"
-
-import * as React from "react"
-import * as ProgressPrimitive from "@radix-ui/react-progress"
-
-const Progress = React.forwardRef<
-  React.ElementRef<typeof ProgressPrimitive.Root>,
-  React.ComponentPropsWithoutRef<typeof ProgressPrimitive.Root>
->(({ className = "", value, ...props }, ref) => (
-  <ProgressPrimitive.Root
-    ref={ref}
-    className={`relative h-4 w-full overflow-hidden rounded-full bg-gray-200 ${className}`}
-    {...props}
-  >
-    <ProgressPrimitive.Indicator
-      className="h-full w-full flex-1 bg-black transition-all"
-      style={{ transform: `translateX(-${100 - (value || 0)}%)` }}
-    />
-  </ProgressPrimitive.Root>
-))
-Progress.displayName = ProgressPrimitive.Root.displayName
-
-export { Progress }
-
